src/test/kotlin/raidhelper/dynamodb/dao/RaidDaoImplTest.kt
package raidhelper.dynamodb.dao

import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
import com.amazonaws.services.dynamodbv2.model.GetItemRequest
import com.amazonaws.services.dynamodbv2.model.GetItemResult
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test
import org.mockito.kotlin.*
import raidhelper.dynamodb.record.RaidRecord
import raidhelper.dynamodb.record.toMap

class RaidDaoImplTest {
    // Setup: Create mocked DynamoDB client, a sample raid, and the result from 'getItem' with necessary type mocking
    val testRaidId = "test-raid-123"
    val mockedDynamoDBClient: AmazonDynamoDB = mock()
    val mockedGetItemResult = mock<GetItemResult>()
    val testRaidRecord = RaidRecord(testRaidId, "Leader", "Date", listOf(), emptyMap(), "status")

    init {
        whenever(mockedDynamoDBClient.getItem(any<GetItemRequest>())).thenReturn(mockedGetItemResult)
        whenever(mockedGetItemResult.item).thenReturn(testRaidRecord.toMap()) // Requires your 'toMap()' extension!
    }

    @Test
    fun `Get Existing Raid Success`() {
        val raidDaoImpl = RaidDaoImpl(mockedDynamoDBClient) // Assuming DI or injection in constructor
        val retrievedRaid = raidDaoImpl.getRaid(testRaidId) // Your DAO instance

        // Assertions
        assertEquals(testRaidId, retrievedRaid?.raidId)
        assertEquals("Leader", retrievedRaid?.raidLeader)
        // Continue similar assertions for other fields/null safety (ideally with more extensive scenarios)
    }
}